Initial commit of openrc scripts
These are modified copies of the sysvinit scripts.

In particular, these are intended for Alpine, so build-packages.sh
unfortunately does not yet work (fpm doesn't support it, so would need
to create new one-off scripts for it, which isn't great).

There's a change to the google_ip_forwarding_daemon Python source to
use `ip route list` instead of `ip route ls` because busybox does not
support `ip route ls`. The output should otherwise be the same across
machines.