Use adduser instead of usermod
usermod is not immediately avaible on alpine and the one available in
the shadow package has a bug around how it handles groups with a large
number of users (http://bugs.alpinelinux.org/issues/7390). This is
intended to address that problem, but may not be the entirely correct
way to do it.